+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - sentry tracing when instrumentation event is not an exception
+- payload deserialization if skip_decoding_error is enabled
 
 ## [0.5.0] - 2023-05-23